refactor(mac): suffix pool files with their platform
The arch library suffixes platform-specific files with the platform they
implement and keeps the header and source names paired, as in
ArchLogUnix.{h,cpp} and ArchDaemonWindows.{h,cpp}. Name the autorelease
pool files the same way.
